Added new macro: AS_UINT8
Although macros obscure code, I chose to add this macro for now because
it reduces verbosity of often-repeated code.
The macro first clamps down a function result to a max value of 255,
and then performs a cast to uint8_t.
The need for the macro arose after I found that i forgot to clamp the output
of the Sobel magnitude result.