Added another macro to convert function values to uint8_t
TO_2D_UINT8_LAMBDA creates a lambda expression which aleviates the
need to a special function just for this conversion.
I am still debating whether these macros are good or bad.  Personally,
I don't like it when code is obfuscated because of macros hiding what's
really going on.  On the other hand, the full code is hard to read and
repeats often.